Move business classes to a subdirectory
Make the source code organization easier to follow by moving all
the monkey business classes to their own directory.  No code
changes, just code reorganization.
